# contracts/public-agent.openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "MyLesson Public Agent API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/ai/course-recommendations": {
      "post": {
        "summary": "Recommend Courses",
        "operationId": "recommend_courses_api_v1_ai_course_recommendations_post",
        "parameters": [
          {
            "name": "x-ml-delegation",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ml-Delegation"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Recommend Courses Api V1 Ai Course Recommendations Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/learning-plan-drafts": {
      "get": {
        "summary": "List Drafts",
        "operationId": "list_drafts_api_v1_ai_learning_plan_drafts_get",
        "parameters": [
          {
            "name": "x-ml-delegation",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ml-Delegation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Drafts Api V1 Ai Learning Plan Drafts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Draft",
        "operationId": "create_draft_api_v1_ai_learning_plan_drafts_post",
        "parameters": [
          {
            "name": "x-ml-delegation",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ml-Delegation"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLearningPlanDraftRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Create Draft Api V1 Ai Learning Plan Drafts Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/learning-plan-drafts/{draft_id}/versions": {
      "get": {
        "summary": "Draft Versions",
        "operationId": "draft_versions_api_v1_ai_learning_plan_drafts__draft_id__versions_get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          },
          {
            "name": "x-ml-delegation",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ml-Delegation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Draft Versions Api V1 Ai Learning Plan Drafts  Draft Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/learning-plan-drafts/{draft_id}/adjustments": {
      "post": {
        "summary": "Adjust Draft",
        "operationId": "adjust_draft_api_v1_ai_learning_plan_drafts__draft_id__adjustments_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          },
          {
            "name": "x-ml-delegation",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ml-Delegation"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Adjust Draft Api V1 Ai Learning Plan Drafts  Draft Id  Adjustments Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/learning-plan-drafts/{draft_id}/confirm": {
      "post": {
        "summary": "Confirm Draft",
        "operationId": "confirm_draft_api_v1_ai_learning_plan_drafts__draft_id__confirm_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          },
          {
            "name": "x-ml-delegation",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ml-Delegation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Confirm Draft Api V1 Ai Learning Plan Drafts  Draft Id  Confirm Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/learning-plan-drafts/{draft_id}/cancel": {
      "post": {
        "summary": "Cancel Draft",
        "operationId": "cancel_draft_api_v1_ai_learning_plan_drafts__draft_id__cancel_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          },
          {
            "name": "x-ml-delegation",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ml-Delegation"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/plans": {
      "get": {
        "summary": "Plans",
        "operationId": "plans_api_v1_ai_plans_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "x-ml-delegation",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ml-Delegation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Plans Api V1 Ai Plans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/plans/{plan_id}/progress": {
      "patch": {
        "summary": "Update Progress",
        "operationId": "update_progress_api_v1_ai_plans__plan_id__progress_patch",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Plan Id"
            }
          },
          {
            "name": "x-ml-delegation",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ml-Delegation"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Update Progress Api V1 Ai Plans  Plan Id  Progress Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdjustmentRequest": {
        "properties": {
          "adjustment": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Adjustment"
          },
          "requestId": {
            "type": "string",
            "format": "uuid",
            "title": "Requestid"
          }
        },
        "type": "object",
        "required": [
          "adjustment",
          "requestId"
        ],
        "title": "AdjustmentRequest"
      },
...
      "CreateLearningPlanDraftRequest": {
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid",
            "title": "Requestid"
          },
          "goal": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Goal"
          },
          "minutesPerDay": {
            "type": "integer",
            "maximum": 480.0,
            "minimum": 10.0,
            "title": "Minutesperday",
            "default": 30
          },
          "courses": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Courses"
          },
          "dailyRoutine": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Dailyroutine"
          }
        },
        "type": "object",
        "required": [
          "requestId",
          "goal",
          "courses",
          "dailyRoutine"
        ],
        "title": "CreateLearningPlanDraftRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProgressRequest": {
        "properties": {
          "progressPercent": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progresspercent"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "progressPercent"
        ],
        "title": "ProgressRequest"
      },
      "RecommendationRequest": {
        "properties": {
          "goal": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Goal"
          },
          "limit": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "goal"
        ],
        "title": "RecommendationRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}